Stuff more or less works.

+++ src/physics/frame_transforms.rs
@@ -16,24 +16,26 @@ fn scale_matrix(v: &Vec2, g: f32) -> Mat3A {
 
 /// Length-contracts every spacetime object relative to player frame.
 pub(crate) fn redraw_in_player_frame(
-    mut q: Query<(&Vel, &mut GlobalTransform), With<SpaceTimeObject>>,
+    mut q: Query<(&Vel, &mut GlobalTransform, Ref<Transform>), With<SpaceTimeObject>>,
 ) {
-    for (Vel(v), mut gt) in q.iter_mut() {
-        if v.length() > 0. {
-            let mut ga = gt.affine();
-            let g = igamma(&v);
-            let a0 = ga.translation.truncate().angle_between(*v);
-            let a1 = f32::acos((f32::cos(a0) - v.length()) / (1. - f32::cos(a0) * v.length()))
-                * a0.signum();
-            ga.translation = Vec2::from_angle(a0 - a1)
-                .rotate(ga.translation.truncate())
-                .extend(ga.translation.z)
-                .into();
-            ga.translation = ga.translation.contract(v);
-            let smat = scale_matrix(v, g);
-            let smat2 = scale_matrix(&v.perp(), 1. - f32::asin(v.length()) * FRAC_2_PI);
-            ga.matrix3 = smat2 * Mat3A::from_rotation_z(a0 - a1) * smat * ga.matrix3;
-            *gt = ga.into();
+    for (Vel(v), mut gt, t) in q.iter_mut() {
+        if t.is_changed() {
+            if v.length() > 0. {
+                let mut ga = gt.affine();
+                let g = igamma(&v);
+                let a0 = ga.translation.truncate().angle_between(*v);
+                let a1 = f32::acos((f32::cos(a0) - v.length()) / (1. - f32::cos(a0) * v.length()))
+                    * a0.signum();
+                ga.translation = Vec2::from_angle(a0 - a1)
+                    .rotate(ga.translation.truncate())
+                    .extend(ga.translation.z)
+                    .into();
+                ga.translation = ga.translation.contract(v);
+                let smat = scale_matrix(v, g);
+                let smat2 = scale_matrix(&v.perp(), 1. - f32::asin(v.length()) * FRAC_2_PI);
+                ga.matrix3 = smat2 * Mat3A::from_rotation_z(a0 - a1) * smat * ga.matrix3;
+                *gt = ga.into();
+            }
         }
     }
 }
